backend/requirements.txt
# django
Django==3.2.8
djangorestframework==3.12.4
django-filter==21.1
django4-background-tasks==1.2.7
django-cors-headers==3.10.0
django-redis==5.0.0
django-extensions==3.1.5
codename==1.1

requests==2.25.1
requests-oauthlib==1.3.0

gunicorn==20.0.4
pyyaml==6.0.1
markdown==3.3.3
uritemplate==3.0.1
coreapi==2.3.3

dateparser==1.0.0
geoip2==4.4.0
chardet==4.0.0
djutils==1.0.13
ua-parser==0.10.0

dnspython==2.3.0
tldextract==5.3.0

django-ipware==7.0.1
humanize==4.13.0
openpyxl==3.1.5
